getNews in DbHElper
