+++ app/src/main/java/com/example/newsfeed/CategoryNewsRVAdapter.java
@@ -34,19 +34,19 @@ public class CategoryNewsRVAdapter extends RecyclerView.Adapter<CategoryNewsRVAd
 
     @Override
     public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
-        ArticleModel articles = articlesArrayList.get(position);
-        holder.subHeadingTV.setText(articles.getDescription());
-        holder.headingTV.setText(articles.getTitle());
-        Picasso.get().load(articles.getUrlToImage()).into(holder.newsIV);
+        ArticleModel article = articlesArrayList.get(position);
+        holder.subHeadingTV.setText(article.getDescription());
+        holder.headingTV.setText(article.getTitle());
+        Picasso.get().load(article.getUrlToImage()).into(holder.newsIV);
         holder.itemView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
                 Intent i = new Intent(context, DetailActivity.class);
-                i.putExtra("title", articles.getTitle());
-                i.putExtra("content", articles.getContent());
-                i.putExtra("desc", articles.getDescription());
-                i.putExtra("imageurl", articles.getUrlToImage());
-                i.putExtra("url", articles.getUrl());
+                i.putExtra("title", article.getTitle());
+                i.putExtra("content", article.getContent());
+                i.putExtra("desc", article.getDescription());
+                i.putExtra("imageurl", article.getUrlToImage());
+                i.putExtra("url", article.getUrl());
                 context.startActivity(i);
             }
         });
